requirements.txt
streamlit
pandas
pillow
python-docx
PyMuPDF
openpyxl
xlrd
python-pptx
google-api-python-client
google-auth
google-cloud-vision
